use std::optional for account companyId

+++ cxx/ui/widget/companieswindow.cpp
@@ -51,9 +51,9 @@ void CompaniesWindow::setCompanies(const QList<domain_id> companyIds) {
 bool CompaniesWindow::confirmDelete(const QSet<const QModelIndex> indexes) {
     QStringList nonEmpty;
     for (auto i : indexes) {
-        if (model.getRow(i)->accounts.toInt() > 0) nonEmpty.append(model.getRow(i)->name.toString());
+        if (model.getRow(i)->accounts > 0) nonEmpty.append(model.getRow(i)->name.toString());
     }
-    // FIXME: delete is disabled for non-empty company
+    // delete is disabled for non-empty company, so the dialog should never be displayed
     return dialog::confirmDelete(this, tr("Confirm delete companies"),
             tr("The following companies have accounts.  "
             "The accounts will remain but will no longer be associated with a company.  "
